+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.01] - 11/20/2022
 
 - Added devotionals
+- Added verse comparison
 
 ## [1.00] - 11/18/2022